test(core): colocate the external thread refetch contract
